Fix sonarcloud bugs

+++ src/MeshKernel/CurvilinearGridFromSplines.cpp
@@ -1659,9 +1659,7 @@ size_t meshkernel::CurvilinearGridFromSplines::MakeGridLine(size_t splineIndex,
         currentMaxWidth = 0.0;
         for (auto n = 0; n < numM; ++n)
         {
-            distances[n] = splineLength * static_cast<double>(n + 1.0) / static_cast<double>(numM);
-
-            //distances[n] = splineLength * (static_cast<double>(n) + 1.0) / static_cast<double>(numM);
+            distances[n] = splineLength * (n + 1.0) / static_cast<double>(numM);
         }
 
         m_splines->InterpolatePointsOnSpline(splineIndex,
